+++ src/tables/graph.rs
@@ -81,7 +81,7 @@ impl<'a> GraphTable<'a> {
         edges: impl Iterator<Item = Edge>,
         workdir: &Path,
         out: &Path,
-        chunk_bytes: u64,
+        chunk_bytes: usize,
     ) -> Result<GraphTable<'a>> {
         let mut writer = Writer::create(out)?;
         let num_edges = AtomicU64::new(0);
@@ -89,7 +89,7 @@ impl<'a> GraphTable<'a> {
             ExternalSorterBuilder::new()
                 .with_tmp_dir(workdir)
                 .with_buffer(LimitedBufferBuilder::new(
-                    chunk_bytes as usize / size_of::<Edge>(),
+                    chunk_bytes / size_of::<Edge>(),
                     /* preallocate */ true,
                 ))
                 .build()?;
@@ -467,7 +467,7 @@ mod tests {
             edges_iter,
             workdir.path(),
             &path,
-            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES as u64,
+            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES,
         )?;
         assert_eq!(graph.modified()?, std::fs::metadata(&path)?.modified()?);
         assert_eq!(
@@ -502,7 +502,7 @@ mod tests {
             edges_iter,
             workdir.path(),
             &path,
-            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES as u64,
+            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES,
         )?;
         assert_eq!(graph.nodes().collect::<Vec<u64>>(), &[1, 2, 3, 4]);
         Ok(())
@@ -518,7 +518,7 @@ mod tests {
             edges_iter,
             workdir.path(),
             &path,
-            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES as u64,
+            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES,
         )?;
         assert_eq!(graph.nodes().collect::<Vec<u64>>(), Vec::<u64>::new());
         Ok(())
@@ -574,7 +574,7 @@ mod tests {
             edges_iter,
             workdir.path(),
             &path,
-            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES as u64,
+            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES,
         )?;
         assert_eq!(graph.edge_count(), 3);
         Ok(())
@@ -589,7 +589,7 @@ mod tests {
             edges_iter,
             workdir.path(),
             &path,
-            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES as u64,
+            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES,
         )?;
         assert_eq!(graph.edge_count(), 0);
         Ok(())
@@ -617,7 +617,7 @@ mod tests {
             edges_iter,
             workdir.path(),
             &path,
-            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES as u64,
+            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES,
         )?;
         // Nodes 1..6 and 21..23, nine in total.
         assert_eq!(graph.node_count(), 9);
@@ -633,7 +633,7 @@ mod tests {
             edges_iter,
             workdir.path(),
             &path,
-            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES as u64,
+            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES,
         )?;
         assert_eq!(graph.node_count(), 0);
         Ok(())
@@ -653,7 +653,7 @@ mod tests {
             edges_iter,
             workdir.path(),
             &path,
-            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES as u64,
+            crate::pipeline::EXTERNAL_SORT_CHUNK_BYTES,
         )?;
         assert_eq!(graph.node_count(), 3);
         Ok(())
